Updated a ton of classes to include the new tips on each page. They are spread about all over the program currently and aren't placed in the greatest of spots but the temporary code is in place so it can be moved wherever. This is all defined in TipsBox and can be repurposed to serve any arraylist of tips one has. Again a future commit should probably replace exactly where these tips are on each page or are in general.
I also made sure the functionality to hide the tips was also in place. I wasn't sure if this was a permanent feature we wanted but it can easily be removed if not. This feature was originally only done in the sports page but now is featured in all 6 spots where the tips appear.